+++ babel.config.js
@@ -14,7 +14,7 @@ module.exports = function(api) {
     api.cache(true);
   }
 
-  const presets = [env(NODE_ENV === 'test' ? 'commonjs' : false), '@babel/preset-react'];
+  const presets = [env(NODE_ENV === 'test' ? 'commonjs' : false), '@babel/preset-typescript', '@babel/preset-react'];
 
   const plugins = [
     '@babel/plugin-proposal-export-default-from',