+++ packages/field/src/Field.tsx
@@ -1,9 +1,10 @@
 /* eslint-disable react/require-default-props */
-import React, { Component } from 'react';
+import React, { PureComponent } from 'react';
 import PropTypes from 'prop-types';
 import isObject from 'lodash/isObject';
 import { hasOwnProperty, classes, noop } from '@swan-form/helpers';
 import asField from './asField';
+import { FieldProps } from './common';
 
 const INPUT_TYPES = [
   'button',
@@ -34,7 +35,7 @@ const INPUT_TYPES = [
 const noWrap = ['button', 'hidden', 'reset', 'submit'];
 
 /* eslint-disable no-use-before-define */
-function renderOption(option) {
+function renderOption(option: any): React.ReactNode {
   // If the option is either a string or a number, then we'll use it for both the key and value.
   if (['string', 'number'].includes(typeof option)) {
     return (
@@ -66,7 +67,7 @@ function renderOption(option) {
   return null;
 }
 
-function renderOptions(options) {
+function renderOptions(options: any | any[]): React.ReactNode {
   if (Array.isArray(options)) {
     // An array can be mapped to the render option method
     return options.map(renderOption);
@@ -102,7 +103,7 @@ function renderOptions(options) {
 }
 /* eslint-enable no-use-before-define */
 
-class Field extends Component {
+class Field extends PureComponent<FieldProps> {
   static displayName = 'Field';
 
   static propTypes = {
@@ -113,8 +114,7 @@ class Field extends Component {
      * An array of error messages (false means no error)
      * @type {Array}
      */
-    errors: PropTypes.arrayOf(PropTypes.oneOfType([PropTypes.string, PropTypes.oneOf([false])]))
-      .isRequired,
+    errors: PropTypes.arrayOf(PropTypes.oneOfType([PropTypes.string, PropTypes.oneOf([false])])).isRequired,
     /**
      * Whether or not the field's value is valid
      * @type {Boolean}
@@ -175,13 +175,9 @@ class Field extends Component {
      * Children of the field
      *
      * @note this applies mostly to usage with buttons
-     * @type {React.element}
+     * @type {React.ReactNode}
      */
-    children: PropTypes.oneOfType([
-      PropTypes.element,
-      PropTypes.string,
-      PropTypes.arrayOf(PropTypes.element),
-    ]),
+    children: PropTypes.oneOfType([PropTypes.element, PropTypes.string, PropTypes.arrayOf(PropTypes.element)]),
     /**
      * The label for the field
      * @type {String}
@@ -201,7 +197,7 @@ class Field extends Component {
     required: PropTypes.bool,
     /**
      * An optional icon component to be merged in with the display
-     * @type {React.Element}
+     * @type {React.ReactNode}
      */
     icon: PropTypes.element,
     /**
@@ -216,10 +212,10 @@ class Field extends Component {
     style: PropTypes.object, // eslint-disable-line
   };
 
-  maybeWrapInLabel(input) {
+  maybeWrapInLabel(input: React.ReactNode) {
     const { className, icon, name, errors, type, label, style } = this.props;
 
-    const spread = {};
+    const spread = {} as { htmlFor?: string; style?: React.CSSProperties };
     if (name) {
       spread.htmlFor = name;
     }
@@ -244,7 +240,7 @@ class Field extends Component {
           {input}
           <span className="sf--icon">{icon && icon}</span>
           <span className="sf--errors">
-            {errors.filter(err => err).map(err => (
+            {errors.filter(Boolean).map((err: string) => (
               <span key={err} className="sf--error">
                 {err}
               </span>
@@ -317,14 +313,7 @@ class Field extends Component {
 
     // Make sure we add a ref and a className to the unwrapped element
     if (type === 'submit' || type === 'reset') {
-      return (
-        <input
-          type={type}
-          ref={setRef}
-          className={classes(['sf--field', className])}
-          {...spreadProps}
-        />
-      );
+      return <input type={type} ref={setRef} className={classes(['sf--field', className])} {...spreadProps} />;
     }
 
     if (type === 'checkbox' || type === 'radio') {
@@ -357,9 +346,7 @@ class Field extends Component {
     }
 
     /* eslint-disable react/no-danger */
-    return (
-      <span dangerouslySetInnerHTML={{ __html: `<!-- Unsupported Field Type (${type}) -->` }} />
-    );
+    return <span dangerouslySetInnerHTML={{ __html: `<!-- Unsupported Field Type (${type}) -->` }} />;
     /* eslint-disable react/no-danger */
   }