+++ packages/field/src/Reset.tsx
@@ -3,7 +3,14 @@ import PropTypes from 'prop-types';
 import isFunction from 'lodash/isFunction';
 import { classes } from '@swan-form/helpers';
 
-export default class Reset extends Component {
+export interface ResetProps {
+  name: string;
+  value: string;
+  resetFunction?(): void;
+  className?: string;
+}
+
+export default class Reset extends Component<ResetProps> {
   static propTypes = {
     name: PropTypes.string,
     value: PropTypes.string,
@@ -21,12 +28,7 @@ export default class Reset extends Component {
     className: '',
   };
 
-  constructor(props) {
-    super(props);
-    this.resetForm = this.resetForm.bind(this);
-  }
-
-  resetForm(event) {
+  resetForm = (event: React.FormEvent) => {
     event.preventDefault();
     event.stopPropagation();
     if (isFunction(this.props.resetFunction)) {
@@ -35,12 +37,10 @@ export default class Reset extends Component {
       this.context.reset();
     } else {
       /* eslint-disable no-console */
-      console.error(
-        'There was no reset function supplied to the reset input on either props or context.',
-      );
+      console.error('There was no reset function supplied to the reset input on either props or context.');
       /* eslint-enable no-console */
     }
-  }
+  };
 
   render() {
     const { className, name, value } = this.props;