+++ packages/helpers/src/hasErrors.js
@@ -1,1 +1,0 @@
-export default errors => errors.filter(x => x).length !== 0;