+++ packages/helpers/src/hasErrors.ts
@@ -1,0 +1,1 @@
+export default (errors: (string | false)[]) => errors.filter(Boolean).length !== 0;