+++ packages/helpers/src/hasOwnProperty.ts
@@ -1,0 +1,2 @@
+export const hasOwnProperty = (obj: object, prop: string): boolean => Object.prototype.hasOwnProperty.call(obj, prop);
+export default hasOwnProperty;