+++ packages/helpers/src/moveCursor.ts
@@ -1,3 +1,11 @@
+import { hasOwnProperty } from './hasOwnProperty';
+export interface ExtendedHTMLElement extends HTMLElement {
+  createTextRange?(): any;
+  value?: any;
+  selectionStart?: number;
+  selectionEnd?: number;
+}
+
 /**
  * Moves the cursor on fields
  *
@@ -9,13 +17,13 @@
  * @param {Number} position the location of the cursor (-1 means to put the cursor at the end)
  * @return {[type]}    [description]
  */
-export default function moveCursor(el, position = -1) {
+export default function moveCursor(el: ExtendedHTMLElement, position = -1) {
   /* eslint-disable no-param-reassign */
-  if (typeof el.selectionStart === 'number') {
+  if (hasOwnProperty(el, 'selectionStart') && typeof el.selectionStart === 'number') {
     const pos = position > -1 ? position : el.value.length;
     el.selectionStart = pos;
     el.selectionEnd = pos;
-  } else if (typeof el.createTextRange !== 'undefined') {
+  } else if (hasOwnProperty(el, 'createTextRange') && typeof el.createTextRange === 'function') {
     const range = el.createTextRange();
     range.collapse(false);
     range.select();